Flesh out a _lot_ more of the data parsing stuff that I needed.
This implements most of the parsing but none (yet) of the length
checks.  It's enough to verify that indeed I'm doing enough sane
work to correctly poke the hardware into associating and getting
valid looking data.

A lot of raw data however gets spat out; so that's disabled from
output by default.